requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kiteconnect==4.1.0
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
pandas==2.1.4
yfinance==0.2.28
scikit-learn==1.3.2
streamlit==1.28.1
plotly==5.17.0
python-telegram-bot==20.7
schedule==1.2.0
requests==2.31.0
asyncio-mqtt==0.13.0
websocket-client==1.6.4
ta==0.10.2
colorama==0.4.6
rich==13.7.0 